fix: adicionando mais service

+++ src/modules/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 
@@ -7,6 +7,11 @@ export class UserService {
   constructor(private prisma: PrismaService) {}
 
   async create(data: CreateUserDto) {
+    const grupo = await this.prisma.grupo.findUnique({ where: { id: data.grupo } });
+    if (!grupo) {
+      throw new NotFoundException(`Grupo com ID ${data.grupo} não encontrado`);
+    }
+
     return await this.prisma.user.create({
       data: {
         nome: data.nome,
@@ -47,6 +52,18 @@ export class UserService {
   }
 
   async remove(id: string) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    await this.prisma.valores.updateMany({
+      where: { grupoId: user.grupoId },
+      data: {
+        totalUsuarios: {
+          decrement: 1,
+        },
+      },
+    });
     return await this.prisma.user.delete({ where: { id } });
   }
